3. MLE/3. Deploy model/testing/requirements.txt
python-dotenv==1.0.1
httpx==0.27.2
uuid==1.30